Rename riak_error message to riak_server_error to avoid conflict with Riak Error
