+++ src/riak_operation.c
@@ -55,6 +55,7 @@ riak_operation_free(riak_operation **rop_target) {
     }
     riak_binary_free(cfg, &(rop->request.bucket));
     riak_binary_free(cfg, &(rop->request.key));
+    riak_binary_free(cfg, &(rop->request.index));
     riak_free(cfg, rop_target);
 }
